Defer hydrating sets so they're always "live"
